refactor(coordinator): extract the trigger-B gate out of types.rs
The Server Guards file-size check failed on `types.rs`: 54201 bytes against
MAX_BYTES=51200 (the LINE limit was fine at 1121/1500). `types.rs` was already
only 2332 bytes under the ceiling on main, so the trigger-B additions tipped it
over -- a latent trap rather than a new one.

Fixed by extracting, not by adding a fourth `djinn:allow-oversize` marker to
this diff. `STREAK_INTERVENTION_THRESHOLD`, `PriorSessionDisposition`, and
`should_route_cycling_intervention` move verbatim into
`dispatch/cycling_intervention.rs` -- the whole trigger-B arming decision in one
module next to its only production caller in `dispatch/task_dispatch.rs`,
matching how `dispatch/outcome.rs` holds `DispatchOutcome`. Their two unit tests
move with them. `types.rs` re-exports all three names, so every `use super::*`
call site is unchanged.

types.rs: 54201 -> 45571 bytes (5629 under the ceiling, and 3297 below main's
own 48868), 1121 -> 947 lines. `scripts/check-file-size.sh` reports 0 oversized
files for this branch's changed set.

Purely mechanical: no behavior change. `cargo test -p djinn-coordinator --lib --
tests::intervention:: dispatch::cycling_intervention types::cooldown_tests` is
82 passed / 0 failed, and re-neutralizing the gate clause in its new home still
fails `cancelled_sessions_do_not_route_the_cycling_planner_intervention` while
the concluded-session test stays green -- the fail-closed semantics survived the
move.
